TITLE/requirements/main.txt
numpy>=1.18.5
scikit_image==0.18.3
pytest>=5.4.3
filelock>=3.0.12
gym==0.19.0
tensorboardX==2.4
imageio-ffmpeg>=0.4.2
jupyter
nbsphinx>=0.7.1
sphinx==4.1.2
sphinxcontrib-napoleon>=0.7
sphinxcontrib-osexample>=0.1.1
sphinx-rtd-theme>=0.5.0
opencv-python
torchmetrics
einops
